Refactor: OpenObserve (#10279)

+++ ct/openobserve.sh
@@ -27,13 +27,16 @@ function update_script() {
     msg_error "No ${APP} Installation Found!"
     exit
   fi
-  msg_info "Updating $APP"
-  systemctl stop openobserve
-  LATEST=$(curl -fsSL https://api.github.com/repos/openobserve/openobserve/releases/latest | grep '"tag_name":' | cut -d'"' -f4)
-  $STD tar zxvf <(curl -fsSL https://downloads.openobserve.ai/releases/openobserve/$LATEST/openobserve-$LATEST-linux-amd64.tar.gz) -C /opt/openobserve
-  systemctl start openobserve
-  msg_ok "Updated $APP"
-  msg_ok "Updated successfully!"
+
+  if check_for_gh_release "openobserve" "openobserve/openobserve"; then
+    msg_info "Updating OpenObserve"
+    systemctl stop openobserve
+    RELEASE=$(get_latest_github_release "openobserve/openobserve")
+    tar zxf <(curl -fsSL https://downloads.openobserve.ai/releases/openobserve/v$RELEASE/openobserve-v$RELEASE-linux-amd64.tar.gz) -C /opt/openobserve
+    systemctl start openobserve
+    msg_ok "Updated OpenObserve"
+    msg_ok "Updated successfully!"
+  fi
   exit
 }
